requirements.txt
azure-functions
azure-data-tables
azure-storage-blob
pillow